+++ hack10/src/main/java/org/hack10/entities/TradingOutpost.java
@@ -19,7 +19,7 @@ public class TradingOutpost extends Entity {
     }
     public void setPosition(Position position) {
         this.position = position;
-        this.hitbox.setLocation(position.x, position.y);
+        this.hitbox.setLocation((int) position.x, (int) position.y);
     }
 
     public void setImageAndHitbox(String filePath) {
